Create static grid with n squares inside
This is the core of the program. It creates a 500x500 px container with small squares
inside evenly spaced, the number of squares per side can be adjusted with the variable 'a'.
When 'a' changes, the size of the container doesn't, while the size of the small squares do
to always perfectly fit the container.