Add 10% darker at each step capability
The solution was to add a the counter 'j' inside the mouseenter event
listener instead of the loop that generates squares.
Essentially, j counts the steps. With this variable it's easy to set
the filter brightness to decrease a 10% at each step. I also used an
if statement (? shorthand) that makes sure that the value
doesn't go below 0%, although I don't know if this was strictly
necessary because the program was working without it too.