Add reset button
The reset button now works well, eliminating all paintings without
changing the grid size. The code is probably not the prettiest, as I
had to add the reset button inside the generateGrid function. Ideally,
I would like it to be it's own function outside of it, but the problem
is that it needs access to the current grid size and I haven't
found a way to do that outside of the function.