Add rainbow toggle
Instead of having a black and a rainbow buttons, the program now paints
black by default and there's a rainbow button toggle. The button can be
a little glichy if you do certain things with it, but overall works
well. Now the functional core of the program is finished
and it is time for styling.