Style the page and improve README
The page has been heavily styled from the previous commit and looks
much closer to the final version. A custom pixel font has been added,
the buttons and slider have been moved to be on the same line above
the grid. The buttons have been styled both on normal and on
click/hover. The slider can still be further styled.